Fix uneven spacing on both sides of the double colon in type declarations
When rendered in browsers, the amount of spacing is uneven on both sides of
the double colon in type declarations. This is due to the use of `&nbsp;`
and the trailing newline character.

On the LHS, there is only one `&nbsp;`, which is rendered as one whitespace.
However, on the RHS, there is one `&nbsp;` as well as a trailing newline
character, which is rendered as two whitespaces.

Avoid the use of `&nbsp;` and adjust Jinja whitespace control to fix
this issue.